+++ src/util/helpers.ts
@@ -1,13 +1,18 @@
 import { HandlerContext } from "@atomist/automation-client";
 import { logger } from "@atomist/automation-client/internal/util/logger";
-import { channel, emoji, escape, url, user } from "@atomist/slack-messages/SlackMessages";
+import * as slack from "@atomist/slack-messages/SlackMessages";
 import * as _ from "lodash";
 import { DirectMessagePreferences } from "../handlers/event/preferences";
 import * as graphql from "../typings/types";
 
+/**
+ * Safely truncate the first line of a commit message to 50 characters
+ * or less.  Only count printable characters, i.e., not link URLs or
+ * markup.
+ */
 export function truncateCommitMessage(message: string, repo: any): string {
     const title = message.split("\n")[0];
-    const escapedTitle = escape(title);
+    const escapedTitle = slack.escape(title);
     const linkedTitle = linkIssues(escapedTitle, repo);
 
     if (linkedTitle.length <= 50) {
@@ -95,7 +100,7 @@ export function repoUrl(repo: any): string {
 }
 
 export function repoSlackLink(repo: any): string {
-    return url(repoUrl(repo), repoSlug(repo));
+    return slack.url(repoUrl(repo), repoSlug(repo));
 }
 
 export function userUrl(repo: any, login: string): string {
@@ -156,6 +161,68 @@ export function isGenerated(node: graphql.PullRequestToPullRequestLifecycle.Pull
     return false;
 }
 
+/**
+ * If the URL is of an image, return a Slack message attachment that
+ * will render that image.  Otherwise return null.
+ *
+ * @param url full URL
+ * @return Slack message attachment for image or null
+ */
+function urlToImageAttachment(url: string): slack.Attachment {
+    const imageRegExp = /[^\/]+\.(?:png|jpe?g|gif|bmp)$/i;
+    const imageMatch = imageRegExp.exec(url);
+    if (imageMatch) {
+        const image = imageMatch[0];
+        return {
+            text: image,
+            image_url: url,
+            fallback: image,
+        };
+    } else {
+        return null;
+    }
+}
+
+/**
+ * Find image URLs in a message body, returning an array of Slack
+ * message attachments, one for each image.  It expects the message to
+ * be in Slack message markup.
+ *
+ * @param body message body
+ * @return array of Slack message Attachments with the `image_url` set
+ *         to the URL of the image and the `text` and `fallback` set
+ *         to the image name.
+ */
+export function extractImageUrls(body: string): slack.Attachment[] {
+    const slackLinkRegExp = /<(https?:\/\/.*?)(?:\|.*?)?>/g;
+    // derived from https://stackoverflow.com/a/6927878/5464956
+    // changed to require HTTP
+    // tslint:disable-next-line:max-line-length
+    const urlRegExp = /\b(https?:\/\/(?:[^\s()<>]+|\(([^\s()<>]+|(\([^\s()<>]+\)))*\))+(?:\(([^\s()<>]+|(\([^\s()<>]+\)))*\)|[^\s`!()\[\]{};:'".,<>?«»“”‘’]))/gi;
+    const attachments: slack.Attachment[] = [];
+    const bodyParts = body.split(slackLinkRegExp);
+    for (let i = 0; i < bodyParts.length; i++) {
+        if (i % 2 === 0) {
+            let match: RegExpExecArray;
+            // tslint:disable-next-line:no-conditional-assignment
+            while (match = urlRegExp.exec(bodyParts[i])) {
+                const url = match[0];
+                const attachment = urlToImageAttachment(url);
+                if (attachment) {
+                    attachments.push(attachment);
+                }
+            }
+        } else {
+            const url = bodyParts[i];
+            const attachment = urlToImageAttachment(url);
+            if (attachment) {
+                attachments.push(attachment);
+            }
+        }
+    }
+    return attachments;
+}
+
 export function extractLinkedIssues(body: string,
                                     repo: any,
                                     ignore: string[] = [],
@@ -183,12 +250,12 @@ export function extractLinkedIssues(body: string,
                         result.repo.forEach(rr => {
                             if (rr.issue) {
                                 rr.issue.forEach(i => {
-                                    results.push({type: "issue", result: i});
+                                    results.push({ type: "issue", result: i });
                                 });
                             }
                             if (rr.pullRequest) {
                                 rr.pullRequest.forEach(pr => {
-                                    results.push({type: "pr", result: pr});
+                                    results.push({ type: "pr", result: pr });
                                 });
                             }
                         });
@@ -254,7 +321,7 @@ export function linkIssues(body: string, repo: any): string {
             const iRegExp = new RegExp(`(${iMatchPrefix})${i}(?!\\w)`, "g");
             const iSlug = (i.indexOf("#") === 0) ? `${repo.owner}/${repo.name}${i}` : i;
             const iUrlPath = iSlug.replace("#", "/issues/");
-            const iLink = url(`${baseUrl}/${iUrlPath}`, i);
+            const iLink = slack.url(`${baseUrl}/${iUrlPath}`, i);
             newPart = newPart.replace(iRegExp, `\$1${iLink}`);
         });
         bodyParts[j] = newPart;
@@ -272,7 +339,7 @@ const gitHubUserMatch = "[a-zA-Z\\d]+(?:-[a-zA-Z\\d]+)*";
  *
  * GitHub usernames may only contain alphanumeric characters or single
  * hyphens, cannot begin or end with a hyphen, and must be 1-36
- * characters.  The maximum lenght is not enforced by this regular
+ * characters.  The maximum length is not enforced by this regular
  * expression, but rather in the getGitHubUsers function.  Mentions
  * must be preceded by an `@` symbol and must not be preceded or
  * followed by any word character.
@@ -467,7 +534,7 @@ export function linkGitHubUsers(body: string = "", context: HandlerContext): Pro
             if (notifier) {
                 notifier.forEach(n => {
                     const mentionRegExp = gitHubUserMentionRegExp(n.login);
-                    body = body.replace(mentionRegExp, "$1" + user(n.person.chatId.screenName));
+                    body = body.replace(mentionRegExp, "$1" + slack.user(n.person.chatId.screenName));
                 });
             }
             return body;
@@ -496,7 +563,7 @@ export function replaceChatIdWithGitHubId(body: string = "", ctx: HandlerContext
                             if (result.ChatTeam[0].members && result.ChatTeam[0].members.length > 0
                                 && result.ChatTeam[0].members[0].person) {
                                 if (result.ChatTeam[0].members[0].person.gitHubId
-                                && result.ChatTeam[0].members[0].person.gitHubId.login) {
+                                    && result.ChatTeam[0].members[0].person.gitHubId.login) {
                                     const login = result.ChatTeam[0].members[0].person.gitHubId.login;
                                     body = body.split(`<@${m}>`).join(`@${login}`);
                                 } else if (result.ChatTeam[0].members[0].person.chatId
@@ -534,8 +601,8 @@ export function getChatIds(str: string): string[] {
 
 export function repoAndChannelFooter(repo: any): string {
     const channels = (repo.channels != null && repo.channels.length > 0 ? " - " + repo.channels.map(c =>
-        channel(c.channelId, c.name)).join(" ") : "");
-    return `${url(repoUrl(repo), repoSlug(repo))}${channels}`;
+        slack.channel(c.channelId, c.name)).join(" ") : "");
+    return `${slack.url(repoUrl(repo), repoSlug(repo))}${channels}`;
 }
 
 /**
@@ -565,14 +632,14 @@ export function isDmDisabled(chatId: ChatId, type?: string): boolean {
 
 export function repoAndlabelsAndAssigneesFooter(repo: any, labels: any, assignees: any[]): string {
 
-    let footer = url(repoUrl(repo), `${repo.owner}/${repo.name}`);
+    let footer = slack.url(repoUrl(repo), `${repo.owner}/${repo.name}`);
     if (labels != null && labels.length > 0) {
         footer += " - "
-            + labels.map(l => `${emoji("label")} ${l.name}`).join(" ");
+            + labels.map(l => `${slack.emoji("label")} ${l.name}`).join(" ");
     }
     if (assignees != null && assignees.length > 0) {
         footer += " - " + assignees.map(a =>
-            `${emoji("bust_in_silhouette")} ${a.login}`).join(" ");
+            `${slack.emoji("bust_in_silhouette")} ${a.login}`).join(" ");
     }
     return footer;
 }